Split pipeline tests by scope to avoid PTXAS slowness/OOM
This is the beginning of a history rewrite annotation.
This is a rewrite of a commit originally made in submodule .upstream-tests.
Original hash: 9e650cb07993b3d1e6de357757904fae57919203.

+++ .upstream-tests/test/cuda/pipeline_memcpy_async_thread_scope_thread_8.pass.cpp
@@ -8,11 +8,11 @@
 
 // UNSUPPORTED: pre-sm-70
 
-#include "pipeline_memcpy_async.h"
+#include "pipeline_memcpy_async_thread_scope_thread.h"
 
 int main(int argc, char ** argv)
 {
-    test_select_source<int64_t>();
+    test_select_source<uint8_t>();
 
     return 0;
 }
